+++ resources/views/crud/formfields/text.blade.php
@@ -2,10 +2,14 @@
     <label class="form-label">{{ $prop['label'] }}</label>
 </td>
 <td>
-    <div>
-        <textarea id="{{ $fieldName }}" name="{{ $fieldName }}" class="ckeditor form-control" @if($readonly) readonly @endif>{{
-            old($fieldName) ?? $model->{$fieldName} ?? $prop['default'] ?? ''
-        }}</textarea>
+    <div class="crud-vuesjs">
+        <custom-ckeditor
+            upload="{{ route('crud-policies.upload') }}"
+            id="{{ $fieldName }}"
+            readonly="{{ $readonly ? 'true' : 'false' }}"
+            csrf="{{ csrf_token() }}"
+            name="{{ $fieldName }}"
+            dbtext="{{ old($fieldName) ?? $model->{$fieldName} ?? $prop['default'] ?? '' }}"></custom-ckeditor>
     </div>
 </td>
 @include('crud-policies::crud.modules.ckeditor', [