+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use Kwaadpepper\CrudPolicies\Http\Controllers\AssetsController;
 use Kwaadpepper\CrudPolicies\Http\Controllers\ModelsController;
+use Kwaadpepper\CrudPolicies\Http\Controllers\UploadController;
 
 $middlewareMinimal = [
     \Illuminate\Routing\Middleware\SubstituteBindings::class
@@ -25,7 +26,11 @@ Route::group([
 
 Route::group([
     'as' => 'crud-policies.models.',
-    'middleware' => array_merge(['throttle:250,1'], $middlewareMinimal),
+    'middleware' => array_merge(
+        ['throttle:250,1'],
+        $middlewareMinimal,
+        config('crud-policies.middlewares')
+    ),
     'namespace' => 'Kwaadpepper\CrudPolicies\Http\Controllers'
 ], function () {
     $routePrefix = config('crud-policies.urlPrefix');
@@ -37,5 +42,23 @@ Route::group([
         ->where('colName', '[a-zA-Z0-9_]*')
         ->where('newOrder', '[0-9]*')
         ->name('changeOrder')
+        ->prefix($routePrefix);
+});
+
+Route::group([
+    'as' => 'crud-policies.',
+    'middleware' => array_merge(
+        config('crud-policies.middlewares'),
+        ['throttle:20,1'],
+        $middlewareMinimal,
+        [\Illuminate\Foundation\Http\Middleware\ValidatePostSize::class],
+    ),
+    'namespace' => 'Kwaadpepper\CrudPolicies\Http\Controllers'
+], function () {
+    $routePrefix = config('crud-policies.urlPrefix');
+
+    // Upload image
+    Route::post('upload', [UploadController::class, 'uploadFile'])
+        ->name('upload')
         ->prefix($routePrefix);
 });